Still dealing with FP_73186 fallout
When warming up the assemblies from disk for the compiler, Carbon.Loader was getting reloaded due to a different asm name thus replacing the in-mem runtime type. We now cope with this by adding an AssemblyResolver edge case: Find out our Facepunch's baptism name and make sure the asm is loaded from AppDomain and not from disk. This is OK because Carbon.Loader should never be hot reloaded from disk.